Noted development of simReserve in ecoinfo package
Added note to help doc for simReserve that the function has been updated
and further developed in ecoinfo pacakge.  Initial developments there
were to allow the circle size to be drawn from a distribution, allowing
circles to overlap the bounday (before trimming), and I think the way
excess area is trimmed out.